pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmq"
version = "0.1.0"
description = "Command-line tool for querying LLMs with templated prompts using LangChain"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "argparse>=1.4.0",
    "loguru>=0.7.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "pydantic>=2.0.0",
]

[project.scripts]
llmq = "llmq.main:main"

